to test whether the member var of normalize_hook can run in parallel

+++ mmaction/datasets/pipelines/formating.py
@@ -277,7 +277,7 @@ class FormatShape:
                 to the next transform in pipeline.
         """
         if not isinstance(results['imgs'], np.ndarray):
-            results['imgs'] = np.array(results['imgs'], dtype=np.float32)
+            results['imgs'] = np.array(results['imgs'])
         imgs = results['imgs']
         # [M x H x W x C]
         # M = 1 * N_crops * N_clips * L
